Clicking near an image point on the map causes a marker to be rendered on the point
We retrieve image points near the click point from the WFS service, and then we find the one that is closest to the click point. The calculations involved in finding the distance between points are based on the approximation that the distance from the equator to the north pole is 10^7 meters, and that the earth is a perfect sphere, neither of which are exactly true, but it is close enough for our purposes.