Retrieve nearby images on same road and lane whenever an image is rendered
These will be needed when we are going to "drive" forward or backwards on the road. I ended up using a bounding box, and then filtering the results based on the properties of the image point (VEGKATEGORI, VEGNUMMER, FELTKODE), instead of using a cql_filter to filter on these properties. It is not possible to use bbox and cql_filter simultaneously. It results in an error message stating exactly that.