Fetch nearby image points only when needed
When the user first clicks on an image point, we go and fetch nearby image points in the same road and lane. Now selecting any of those nearby image points does not trigger a new fetch of image points, unless we get close to the end of the list at either end. Then we need to fetch nearby image points again.