Draw a directional marker if the current imagePoint has direction info i metadata
RETNING signifies the directional angle of the image. This can be used directly as rotationAngle of a Leaflet marker by help of the leaflet-rotatedmarker plugin. If the RETNING property is missing, we simply use a non-directional marker icon.

Future improvement: For points which lack the RETNING property, we can infer the direction by computing a vector from the previous point to the next point.