Take into account that FELTKODE (lane code) can contain more than just a single number
FELTKODE is a string, of which the first character is the primary lane number. An odd number here signifies that the lane goes in the direction of increasing meter-values, and even numbers signify the opposite direction. The rest of the lane code describe other attributes of the lane, which are not relevant for the purpose of finding a lane in the opposite direction. (See: https://datakatalogen.vegdata.no/793-NVDB%20dokumentasjon)