Adjust filter for nearby images in current lane
When fetching nearby images for the current road and lane (which we expect we will soon need to display when driving forwards or backwards), we need to filter out images belonging to a different kryssdel or sideanleggsdel. Failing to do this may lead to images from ramps and the like being mixed in with the images of the regular lane.

A side effect of this new filter is that roundabouts are completely skipped over when coming from a regular lane, but this is preferable. (Previously they were only skipped if you came from a lane with a different feltkode than that of the roundabout.)