Sort nearby images by STREKNING, DELSTREKNING, METER
This guarantees that they are in the right order. Previously they were sorted by date and time, which makes no such guarantees (but was correct most of the time).

Now the problem is that we don't know which direction is forwards and which is backwards. I will deal with that next.