Draw individual markers based on WFS query at high zoom levels
I am testing out whether drawing individual markers for each image point is feasible. Instead of simply fetching WMS tiles, we can call the WFS service to retrieve all the image points within the visible map area and render them as individual markers. This has the following advantages:
- We can control how the markers are rendered directly within the app
- The markers can be interacted with, and you get a different cursor when pointing at them, indicating interactivity to the user

The downside is that drawing lots of markers reduces performance. I have restricted this feature to be used only at high zoom levels. Also I have conceptually divided the world into "tiles" of constant size (in fractions of a degree of latitude and longitude), so that we can easily cache the responses from the WFS service, although the primary issue is the rendering.

Things to improve:
- Only draw markers inside the visible map area. (We are currently fetching and rendering image points for all conceptual "tiles" which overlap the visible map area. When rendering markers we can prune away those that are outside the map bounds.)
- Use local storage to cache responses from the WFS service
- Remove duplicate image points before rendering. (Some markers from just outside the conceptual tile borders are included in the response from the WFS service, for some reason. When fetching several "tiles", duplicate image points will occur.)
- Draw directional markers

If performance remains a hindrance, we may have to stick with just using the WMS service for rendering image points, and using the WFS service to retrieve the selected point on map click and image view navigation.