Draw markers at high zoom levels and WMS tiles at lower zoom levels
Also remove the SelectedImagePoint component, as now we can set the current image point simply through the onclick property of the marker.

With this implementation, it is not possible to select image points at lower zoom levels (no markers to actually click on), but the WMS tiles enable you to see where there are image points and zoom in on the appropriate area.