Create MapContainer with Leaflet map and a tile layer with background map
The nvdbcache background map is only available with crs EPSG:25833 (UTM zone 33N), which is not supported natively by Leaflet, so the proj4leaflet plugin must be used.

Zoom levels and origin are documented here: https://nvdbcache.geodataonline.no/arcgis/rest/services/Trafikkportalen/GeocacheTrafikkJPG/MapServer

I could not use the makeStyles hook to style the Leaflet map (make the map fill the available space). The Map component from React-Leaflet creates a div with class "leaflet-container", but I could not apply styles to that class using the makeStyles hook (and specifying styles for leafletContainer). So I had to create a regular css file for this.