Show road image in a popup for each marker
Default behavior: The popup appears when you click on a marker, and disappears when you click outside it. It may be better to show the popup when hovering over the marker. Also the styling of the popup needs to be improved. Should probably just show the image. No bubble or frame or anything.