Enable drop shadow below the Header and set correct background color in the image container
Currently using the elevation property of the AppBar to get the shadow effect. If more flexibility in styling the shadow is needed, we could use css box-shadow directly instead.

Had to set z-index of the Header to a high value to bring it in front of the content below, or the shadow would not be visible.