Create separate component representing the image points layer on small map
The behavior of the image points layer on the small map differs quite a bit from that on the large map, so a separate component makes more sense. Currently the ImagePointsLayerSmallMap component is in a sort of debug mode. Various bounding boxes are rendered on the map to help visualize where and when image points are fetched and rendered. in order to actually see all those boxes within the map area, we simulate a smaller visible map area than is actually the case.

The new LoadedImagePointsContext ensures that we already have image points loaded when the user clicks an image point in the large map, and the app switches to the image viewer with small map.