Add state and effect hooks to ImageViewer for keeping track of image points on the current road/lane
We need to know the next and previous image points in order to go forwards and backwards, and to find those, we need a sorted list of image points for the current lane. For this purpose the concept of a road context is introduced, which encompasses VEGKATEGORI, VEGNUMMER, KRYSSDEL, SIDEANLEGGSDEL and FELTKODE. This is used in a filter which extracts the relevant image points from the currently loaded image points. We could have used these properties directly from currentImagePoint, but we do not want the filtering to happen whenever the currentImagePoint changes (which can be fairly often). The filtering should only be done whenever new image points have been loaded, or the road context has changed (user selected an image point in a different lane or on a different road altogether).