Automatically pan the map to the newly selected image point after moving forwards or backwards in the image viewer
Also make a change to the CurrentCoordinatesContext, allowing for setting latlng without also setting the zoom level. Makes it simpler to update coordinates and keep whatever zoom level is currently set.

Also add a utility function for conveniently getting the coordinates of an image point, in the form of a latlng object.