Fix bug with navigating image points on a kryssdel or sideanleggsdel
When filtering the currently loaded images to find those that are in the same lane as the current image point, we previously only looked at VEGKATEGORI, VEGNUMMER, KRYSSDEL, SIDEANLEGGSDEL and FELTKODE. The problem is that when KRYSSDEL or SIDEANLEGGSDEL has a value, that value is usually 1. (This caused wild jumps between different kryssdels when clicking the forward or backward buttons.) I believe it has a different value only when there are several kryssdels or sideanleggsdels with the same anchorpoint on the actual road. So we have to include ANCHORPOINT in the filter. ANCHORPOINT is a meter value, which relates to the current STREKNING and DELSTREKNING, so we include those in the filter as well.