Keep the layout from wrapping when the height of the image exceeds the height of the viewport
For some reason "nowrap" is not default in Material UI Grid, as opposed to the underlying flexbox. Now with wrapping disabled the image pushes the footer lower, causing scrollbars to appear. This is an improvement, but I still have to find a way to keep the image scaled appropriately between the header and footer.