Do not open full screen image viewer after search
After search we remain in the mode that we were before. The nearest image point to the searched location is selected. This means that the ImagePointLayer component (which is used to fetch and display image points in the big map) needs to process the selectNearestImagePoint command, exactly like the ImageViewer component does.