Create MapControls component with a custom ZoomControl inside
Also disable the native Leaflet zoom control.

The custom ZoomControl component uses the useLeaflet hook to access the map object, on which zoomIn() and zoomOut() can be called. The zoom property of the CurrentCoordinatesContext is already being updated on each viewport-change. The code for this is in the onViewPortChanged property of the Map component (in MapContainer and SmallMapContainer).