Add a splash screen with a welcome message and information
For now it contains placeholder text. The splash screen is visible when the app loads, and you can close it with the close button, after which it will not be shown again until you reload the page.

Next I need to make it close when the user clicks outside it, and also allow the user to tick a box to not be shown the message again. Will have to use local storage to save that preference.