Inital work for supporting the selection of a specific image series for the current road context
At the outset, the newest series for each road context is included as before. I have now added ImageSeriesContext, which enables the selection of a specific image series (ie. an earlier date in the current year) for the current road context (vegkategori, vegstatus, vegnummer, strekning, delstrekning, kryssdel, sideanleggsdel). If a specific series is selected, then that series is included instead of the newest for the current road context.

A missing step is to make the ImageViewer make use of the resulting filtered set of image points instead of the unfiltered set (which includes imagepoints from all series).