Large overhaul of the information flow for handling image series
Previously there was a lot of logic related to updating current and available image series in odd places (CurrentImagePointContext and ImagePointsLayer). With the new structure the information flows naturally from the outermost contexts towards the innermost ones, updating state as necessary. There is now less state updating going on the actual React components.