Automatically select the image point which is closest to the current one after switching image series or year
When swithing between years, no image point will be selected if the road has changed VEGKATEGORI, VEGSTATUS or VEGNUMMER between the two years. If only the FELTKODE changed, it could be that an image point in another lane is selected (for example if the metering direction of the road was changed).

Also fixed a recently introduced bug in Search, where no image point would be selected automatically after the jump. (Had to reset the newly added filteredImagePoints.)