Show info message when attempting to drive further than the end of the image series
If you attempt to go forward when there is no next image (happens at the end of the road or current image series), you get a snackbar (also called toast) with a message about this.

Same thing happens if you try to go backwards when there is no previous image (at the first image of the road or series).