Refactor the switching between map view and image view and reflect which view is active in a query parameter in the URL
We need the URL to mirror some of the application's state in order to support sharing of URLs to open the app in a specific state. This is a beginning of that.

The useQueryParamState hook is courtesy of Kristoffer Giltvedt Selbekk. Really an elegant solution. I was mucking about with react-router without creating anything but problems. This is so much cleaner.

Also refactored App.js a bit. Extracted all the actual React components into ComponentsWrapper.