Reflect current image id in URL and initialize current image point based on image id from URL if present
A known weakness with this approach: If you pan far enough away or zoom out so that loadedImagePoints no longer contains the image point with the current image id, and you copy the URL and paste it in a browser, no image point will be selected, since no image point with the requested id can be found. A fix for this might be to set the currentImagePoint to null as soon as it is no longer available in loadedImagePoints, but this might also be disturbing for users.

Also need to check that the image id query param is valid before using it. Will do that next.