Fix bug: Not possible to turn around when there are no images in lane 1
The code for switching directions naively assumed that there is always a lane with FELTKODE 1 and 2 (or at least having 1 as the first character). Actually that should be correct (on roads with opposite lanes), but it is possible that there are no image points in that lane. In such places it was not possible to switch from lane 2,4,... to an opposite lane, even though there were images in lane 3.

Fix this by simply comparing the parity instead of looking for lane 1 or 2.