Reduce size of conceptual tiles used for fetching from WFS for possible performance improvement
Reducing the size of the tiles means that more requests will be performed, but the browser does them in parallell. This appears to improve performance somewhat in the app, but it will put a heavier load on the WFS service (albeit for a shorter amount of time).

The reason for trying to improve performance now is that the WFS service has become quite a bit slower after metadata for drawing the meter line was included for each image point. If performance remains an issue because of this, we should look at taking that data out again, and instead create a separate web service for getting that data for a single image at a time.