Copy URL to clipboard when clicking the share button
The URL is taken from window.location, and the coordinates query param is replaced with the coordinates of the current image point (zoom level set to 16), in case the user dragged the map away from the selected image point before clicking the share button. (See comment in urlUtilities.js.)

There is currently no feedback informing the user that the URL was copied to the clipboard. Will do that next.