Match search string against regex for vegsystemreferanse before performing search
This is primarily to avoid calling the NVDB API with misformed queries.

If the search string matches a complete or partial vegsystemreferanse, it is formatted in a consistent manner and padded with the missing parts. Example: If you search for "EV6", the query will be padded out to "EV6 S1D1 m1". If you search for "EV6 S12D1", the query will be padded out to "EV6 S12D1 m1".