Use ref hook to access properties of the image in ImageViewer
Will need to use this in order to scale and position the canvas correctly.

The imageLoaded state was necessary to hold off on setting the imageElement variable until after the image has actually loaded in the DOM. Otherwise, it would be undefined when loading the image view directly (from URL with query param view=image).