Slightly modified Map-disabled and Measure-disabled icons
These icons included a line in a different color than the rest of the icon. This was a sort of shadow line just behind the regularly colored line. Seeing as we need to be able to change the icon color to orange on hover, this was problematic, as it is not possible (to my knowledge) to target individual paths in the SVG with css styling.