Adapt useQueryParamState to handle query parameters which are supposed to be integers
useQueryParamState can now optionally treat the query param as an integer, ie. parse it as such before saving the initial state from query param. It will also verify that the new state is a number when setting state.

This fixes a problem with initially rendering the wrong markers for image points, stemming from the comparison of the string "2020" and the number 2020 in ImagePointsLayer.getMarkerIcon (line 191).